+++ packages/vue/src/components/SynVueRadioGroup.vue
@@ -77,6 +77,17 @@ instead.
   size?: SynRadioGroup['size'];
 
   /**
+* The layout of the radio group.
+* This determines how the radios are displayed.
+This only applies to `syn-radio` elements.
+`syn-radio-button` items are rendered in their own button-group and will always render vertically.
+
+- `horizontal`: Radios are displayed in a row.
+- `vertical`: Radios are displayed in a column.
+ */
+  layout?: SynRadioGroup['layout'];
+
+  /**
 * By default, form controls are associated with the nearest containing `<form>` element.
 * This attribute allows you
 to place the form control outside of a form and associate it with the form that has this `id`.